UI_Utility: move some utility funcs to new file
