+++ dllmain/cfgMenu.cpp
@@ -10,6 +10,7 @@
 #include <FAhashes.h>
 #include "Utils.h"
 #include "UI_DebugWindows.h"
+#include "UI_Utility.h"
 #include "Trainer.h"
 
 bool bCfgMenuOpen;
@@ -25,23 +26,6 @@ float fLaserColorPicker[3]{ 0.0f, 0.0f, 0.0f };
 
 ImVec2 PrevRightColumnPos;
 ImVec2 PrevLeftColumnPos;
-
-enum class MenuTab
-{
-	Display,
-	Audio,
-	Mouse,
-	Keyboard,
-	Controller,
-	Framerate,
-	Misc,
-	Memory,
-	Hotkeys,
-	Trainer,
-	NumTabs,
-};
-
-MenuTab Tab = MenuTab::Display;
 
 int MenuTipTimer = 500; // cfgMenu tooltip timer
 
@@ -62,144 +46,6 @@ bool ParseConfigMenuKeyCombo(std::string_view in_combo)
 	}, &cfgMenuCombo });
 
 	return cfgMenuCombo.size() > 0;
-}
-
-// If we don't run these in another thread, we end up locking the rendering
-void SetHotkeyComboThread(std::string* cfgHotkey)
-{
-	bWaitingForHotkey = true;
-
-	pInput->set_hotkey(cfgHotkey, true);
-
-	bWaitingForHotkey = false;
-	return;
-}
-
-void SetHotkeyThread(std::string* cfgHotkey)
-{
-	bWaitingForHotkey = true;
-
-	pInput->set_hotkey(cfgHotkey, false);
-
-	bWaitingForHotkey = false;
-	return;
-}
-
-bool ImGui_TabButton(const char* btnID, const char* text, const ImVec4 &activeCol, 
-	const ImVec4 &inactiveCol, MenuTab tabID, const char* icon, const ImColor iconColor,
-	const ImColor textColor, const ImVec2 &size = ImVec2(0, 0))
-{
-	ImDrawList* drawList = ImGui::GetWindowDrawList();
-
-	ImGui::PushStyleColor(ImGuiCol_Button, Tab == tabID ? activeCol : inactiveCol);
-	bool ret = ImGui::Button(btnID, ImVec2(172, 31));
-	ImGui::PopStyleColor();
-
-	auto p0 = ImGui::GetItemRectMin();
-	auto p1 = ImGui::GetItemRectMax();
-
-	float textHeight = ImGui::CalcTextSize(text).y;
-
-	float y_text_offset = (size.y - textHeight) * 0.5f;
-	float y_icon_offset = y_text_offset + 2.0f;
-
-	drawList->AddText(ImGui::GetFont(), ImGui::GetFontSize(), ImVec2(p0.x + 35.0f, p0.y + y_icon_offset), iconColor, icon, NULL, 0.0f, &ImVec4(p0.x, p0.y, p1.x, p1.y));
-	drawList->AddText(ImGui::GetFont(), ImGui::GetFontSize(), ImVec2(p0.x + 65.0f, p0.y + y_text_offset), IM_COL32_WHITE, text, NULL, 0.0f, &ImVec4(p0.x, p0.y, p1.x, p1.y));
-
-	if (ret && tabID != MenuTab::NumTabs)
-		Tab = tabID;
-
-	return ret;
-}
-
-void ImGui_ItemBG(float RowSize, ImColor bgCol)
-{
-	// Works best when used inside a table
-
-	ImDrawList * drawList = ImGui::GetWindowDrawList();
-	ImVec2 p0 = ImGui::GetCursorScreenPos();
-	float top = ImGui::GetCursorPosY();
-
-	drawList->AddRectFilled(ImVec2(p0.x - 10, p0.y - 10), ImVec2(p0.x + 10 + ImGui::GetContentRegionAvail().x, p0.y + RowSize - top + 7), bgCol, 5.f);
-}
-
-void ImGui_ItemSeparator()
-{
-	// ImGui::Separator() doesn't work inside tables?
-
-	ImDrawList * drawList = ImGui::GetWindowDrawList();
-	ImVec2 p0 = ImGui::GetCursorScreenPos();
-
-	drawList->AddLine(ImVec2(p0.x, p0.y), ImVec2(p0.x + ImGui::GetContentRegionAvail().x, p0.y), ImGui::GetColorU32(ImGuiCol_Separator));
-}
-
-void ImGui_ItemSeparator2()
-{
-	// ImGui::Separator() doesn't work inside tables?
-
-	ImDrawList* drawList = ImGui::GetWindowDrawList();
-	ImVec2 p0 = ImGui::GetCursorScreenPos();
-
-	drawList->AddRectFilled(ImVec2(p0.x, p0.y), ImVec2(p0.x + ImGui::GetContentRegionAvail().x, p0.y + 3), ImColor(150, 10, 40));
-}
-
-ImColor itmbgColor = ImColor(25, 20, 20, 166);
-int itemIdx = -1;
-float columnLastY[2] = { 0,0 };
-void ImGui_ColumnInit()
-{
-	itemIdx = -1;
-	columnLastY[0] = columnLastY[1] = 0;
-
-	ImGui::TableNextColumn();
-}
-
-std::vector<float> bgHeightsPerTab[int(MenuTab::NumTabs)]; // item heights, per-tab
-void ImGui_ColumnSwitch()
-{
-	auto& bgHeights = bgHeightsPerTab[int(Tab)];
-	if (itemIdx >= 0)
-	{
-		// Store bgheight of previous item, so we can use it as a BG height.
-		// (ImGui really should have a way to simply get the current row/cell height of a table instead...
-		if (bgHeights.size() > size_t(itemIdx))
-			bgHeights[itemIdx] = ImGui::GetCursorPos().y;
-		else
-			bgHeights.push_back(ImGui::GetCursorPos().y);
-
-		ImGui::Dummy(ImVec2(10, 25));
-		columnLastY[itemIdx % 2] = ImGui::GetCursorPos().y;
-	}
-
-	itemIdx++;
-	ImGui::TableSetColumnIndex(itemIdx % 2);
-
-	if (itemIdx > 1) // 0/1 are first item on column, no need to restore lastY
-		ImGui::SetCursorPosY(columnLastY[itemIdx % 2]);
-
-	float bgHeight = 0;
-	if (bgHeights.size() > size_t(itemIdx))
-		bgHeight = bgHeights[itemIdx];
-	else
-		bgHeight = 0;
-
-	ImGui_ItemBG(bgHeight, itmbgColor);
-}
-
-void ImGui_ColumnFinish()
-{
-	if (itemIdx >= 0)
-	{
-		auto& bgHeights = bgHeightsPerTab[int(Tab)];
-
-		// Store bgheight of previous item
-		if (bgHeights.size() > size_t(itemIdx))
-			bgHeights[itemIdx] = ImGui::GetCursorPos().y;
-		else
-			bgHeights.push_back(ImGui::GetCursorPos().y);
-	}
-
-	ImGui::Dummy(ImVec2(10, 25));
 }
 
 void cfgMenuRender()
