+++ dllmain/dllmain.vcxproj
@@ -160,6 +160,7 @@ echo | set /p dummyName=#define GIT_BRANCH ^"%VAR%^" &gt;&gt; gitparams.h</Comma
     <ClCompile Include="AspectRatioTweaks.cpp" />
     <ClCompile Include="Trainer.cpp" />
     <ClCompile Include="UI_DebugWindows.cpp" />
+    <ClCompile Include="UI_Utility.cpp" />
     <ClCompile Include="Utils.cpp" />
     <ClCompile Include="WndProcHook.cpp" />
     <ClCompile Include="ToolMenu.cpp" />
@@ -269,6 +270,7 @@ echo | set /p dummyName=#define GIT_BRANCH ^"%VAR%^" &gt;&gt; gitparams.h</Comma
     <ClInclude Include="resource.h" />
     <ClInclude Include="Trainer.h" />
     <ClInclude Include="UI_DebugWindows.h" />
+    <ClInclude Include="UI_Utility.h" />
     <ClInclude Include="Utils.h" />
   </ItemGroup>
   <ItemGroup>
